backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
pandas==2.1.2
google-generativeai==0.8.2
python-dotenv==1.0.1